LayoutTests/fast/scrolling/scroll-anchoring-with-page-scale-expected.txt

PASS window.pageYOffset is 400
PASS successfullyParsed is true

TEST COMPLETE
